+++ src/App.tsx
@@ -1,12 +1,50 @@
 import { useEffect, useState } from "react";
 import { FeatureList } from "./components/FeatureList";
-import { Link } from "react-router-dom";
-import { getTokenFromCookie, spotifyFetch } from "./utils/constants";
+import {
+  getTokenFromCookie,
+  setTokenCookie,
+  spotifyFetch,
+} from "./utils/constants";
 import "./styles/global.css";
+
+const clientId = process.env.REACT_APP_CLIENT_ID;
+const redirectUri = process.env.REACT_APP_REDIRECT_URI;
+
+function objectToQueryString(obj: any) {
+  return Object.entries(obj)
+    .map(([key, value]) => `${key}=${value}`)
+    .join("&");
+}
+
+function getAuthUrl() {
+  const spotifyAuthEndpoint = "https://accounts.spotify.com/authorize?";
+
+  if (!clientId || !redirectUri) {
+    throw new Error("Missing config");
+  }
+
+  const params = {
+    client_id: clientId,
+    response_type: "code",
+    redirect_uri: redirectUri,
+    scope: "user-library-modify user-library-read user-read-private",
+  };
+
+  const queryString = objectToQueryString(params);
+
+  return spotifyAuthEndpoint + queryString;
+}
 
 function App() {
   const [user, setUser] = useState();
-  const token = getTokenFromCookie();
+  // sort this shit out
+  const token =
+    getTokenFromCookie() ||
+    new URLSearchParams(window.location.search).get("token");
+
+  if (token) {
+    setTokenCookie(token);
+  }
 
   useEffect(() => {
     const fetchUser = async () => {
@@ -25,7 +63,7 @@ function App() {
   return (
     <>
       <h1>Spotify Tools</h1>
-      {!token && <Link to="/login">Login</Link>}
+      {!token && <a href={getAuthUrl()}>Login</a>}
       {user && <p>Hi, {user}!</p>}
       <FeatureList />
     </>
