+++ src/index.tsx
@@ -2,7 +2,6 @@ import React from "react";
 import ReactDOM from "react-dom";
 import App from "./App";
 import { BrowserRouter, Route, Routes } from "react-router-dom";
-import Login from "./Login";
 import ExportImport from "./routes/ExportImport";
 
 ReactDOM.render(
@@ -10,7 +9,6 @@ ReactDOM.render(
     <BrowserRouter>
       <Routes>
         <Route path="/" element={<App />} />
-        <Route path="login" element={<Login />} />
         <Route path="export-import" element={<ExportImport />} />
       </Routes>
     </BrowserRouter>
